# src/app/api/v2/doc/swagger.yml
openapi: "3.0.0"
servers:
  - description: dev server
    url: http://localhost:3001/api/v2
  - description: live server
    url: http://rbs-erp-next.vercel.app/v1/
  - description: googl server
    url: https://rbs-erp-next.vercel.app/v1/
info:
  description: RBS Inventory API
  version: "2.0.0"
  title: RBS Inventory API
  contact:
    email: rakibulhasan.bdm@gmail.com
    name: Rakibul Hasan
  license:
    name: Apache 2.0
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
...
paths:
...
  "/accounts":
    get:
      description: get customers
      security:
        - bearerAuth: []
      tags:
        - "Accounts-routes"
      parameters:
        - $ref: "#/components/parameters/pageParams"
        - $ref: "#/components/parameters/limitParams"
        - $ref: "#/components/parameters/sortTypeParams"
        - $ref: "#/components/parameters/sortKey"
        - $ref: "#/components/parameters/search"
        - $ref: "#/components/parameters/searchBy"
      responses:
        "200":
          description: Successfully get customers
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Ok
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/accounts"
                  pagination:
                    $ref: "#/components/schemas/pagination"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
    post:
      description: create chart of account
      security:
        - bearerAuth: []
      tags:
        - Accounts-routes
      requestBody:
        content:
          "application/json":
            schema:
              $ref: "#/components/schemas/accounts"
      responses:
        "201":
          description: Successfully create account
          content:
            "application/json":
              schema:
                type: object
                properties:
                  code:
                    type: integer
                    format: int32
                    example: 200
                  message:
                    type: string
                    example: Ok
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/accounts"
        "409":
          $ref: "#/components/responses/409"
        "400":
          $ref: "#/components/responses/400"
        "500":
          $ref: "#/components/responses/500"
...
components:
  schemas:
...
    accounts:
      properties:
        _id:
          type: string
          example: 634e372366d40b495f02d980
        account_name:
          type: string
          example: abcs
        account_type:
          type: string
          enum: ["Expense", "Income", "Liability", "Asset"]
        is_system_account:
          type: boolean
        status:
          type: string
          enum: ["active", "inactive"]
        description:
          type: string
        is_debit:
          type: boolean
        created_by:
          type: string
    pagination:
      type: object
      properties:
        nextPage:
          type: integer
          example: 3
        prevPage:
          type: integer
          example: 1
        currentPage:
          type: integer
          example: 2
        totalPage:
          type: integer
          example: 5
  responses:
    "500":
      description: Internal server error
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                format: int32
                example: 500
              message:
                type: string
                example: We are sorry for this convenience, Please try again later
              data:
                example: any
            required:
              - "code"
              - "message"
    "400":
      description: Bad Request
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 400
              message:
                type: string
                example: "Bad Request"
              data:
                type: object
                example:
                  field: "field is required"
                  field2: "field2 is required"

            required: [code, message]
...
    "401":
      description: Unauthorized
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 401
              message:
                type: string
                example: Access denied
...
    "409":
      description: Conflict
      content:
        "application/json":
          schema:
            type: object
            properties:
              code:
                type: integer
                example: 409
              message:
                type: string
                example: Conflict with account_name
              data:
                type: object
                example:
                  field: "Conflict with this name"
  parameters:
...
    pageParams:
      in: query
      name: page
      description: current page number
      schema:
        type: integer
        example: 1
      required: true
...
    limitParams:
      in: query
      name: limit
      description: maximum items to be returned
      schema:
        type: integer
        example: 5
    sortTypeParams:
      in: query
      name: sort_type
      description: the type of sort (asc or desc)
      schema:
        type: string
        enum:
          - asc
          - desc
    sortKey:
      in: query
      name: sort_key
      description: sort_key
      schema:
        type: string
        example: createdAt
    search:
      in: query
      name: search
      description: search terms
      schema:
        type: string
        example: ""
    searchBy:
      in: query
      name: search_by
      description: search field name
      schema:
        type: string
        example: "key"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT